TDAmeritrade.jl

Documentation for TDAmeritrade.jl

Getting started

Start off by running TD_auth() and follow the instruction.

APIs

TDAmeritrade.TD_auth — Function

TD_auth()

run authentication procedure

already authenticated -> do nothing

access token expired && has refresh token -> refresh access token

refresh token expired -> get refresh token

no CONSUMER_KEY -> error

TDAmeritrade.price_history — Function

price_history(ticker; kwargs...)

Get price history of a given ticker, everything but ticker have default value specified by TD api: https://developer.tdameritrade.com/price-history/apis/get/marketdata/{symbol}/pricehistory

optional kwargs: periodType period frequencyType frequency endDate startDate needExtendedHoursData

price_history(ticker, freq, peri)

see TD API for default and compatible frequency vs. tickers

price_history(ticker, freq, start::DateTime=now()-Day(1), stop::DateTime=now())

get the price history with interval ticks from start to stop

valid frequency: minute: 1*, 5, 10, 15, 30

TDAmeritrade.get_quotes — Function

get_quotes(ticker)

Get quote of a symbol: https://developer.tdameritrade.com/quotes/apis/get/marketdata/{symbol}/quotes

required: ticker

get_quotes(tickers::Array)

get multiple quotes at once